server/requirements.txt
polars
scikit-learn
matplotlib
pandas
numpy
seaborn
python-dotenv
